+++ bob-common/src/metrics/mod.rs
@@ -4,7 +4,6 @@ use crate::{
 };
 use metrics::{register_counter, Recorder};
 use metrics_exporter_prometheus::PrometheusRecorder;
-use metrics_util::MetricKindMask;
 use pearl::init_pearl;
 use std::{
     net::SocketAddr,
@@ -214,13 +213,13 @@ fn init_grinder() {
 }
 
 fn init_backend() {
-    register_counter!(BACKEND_STATE);
-    register_counter!(BLOBS_COUNT);
-    register_counter!(ALIEN_BLOBS_COUNT);
+    register_gauge!(BACKEND_STATE);
+    register_gauge!(BLOBS_COUNT);
+    register_gauge!(ALIEN_BLOBS_COUNT);
 }
 
 fn init_link_manager() {
-    register_counter!(AVAILABLE_NODES_COUNT);
+    register_gauge!(AVAILABLE_NODES_COUNT);
 }
 
 async fn install_global(node_config: &NodeConfig, local_address: &str) -> SharedMetricsSnapshot {
@@ -233,6 +232,9 @@ async fn install_global(node_config: &NodeConfig, local_address: &str) -> Shared
     if node_config.metrics().prometheus_enabled() {
         let prometheus_rec = build_prometheus(node_config);
         recorders.push(Box::new(prometheus_rec));
+        info!("prometheus exporter enabled");
+    } else {
+        info!("prometheus exporter disabled");
     }
 
     if !recorders.is_empty() {
@@ -260,10 +262,10 @@ fn build_prometheus(node_config: &NodeConfig) -> PrometheusRecorder {
         .expect("Bad prometheus address");
     let (recorder, exporter) = metrics_exporter_prometheus::PrometheusBuilder::new()
         .listen_address(addr)
-        .idle_timeout(MetricKindMask::ALL, Some(Duration::from_secs(2)))
         .build_with_exporter()
         .expect("Failed to set Prometheus exporter");
 
+    debug!("prometheus built");
     let future = async move {
         pin!(exporter);
         loop {